Time: 110 ms (60.00%) | Memory: 84.8 MB (71.73%) - LeetSync

+++ 2049-eliminate-maximum-number-of-monsters/eliminate-maximum-number-of-monsters.cpp
@@ -4,7 +4,7 @@ public:
         int n = dist.size();
         vector<double> time(n);
         for(int i = 0; i<n; i++){
-            time[i] = ((double)(dist[i]))/((double)(speed[i]));
+            time[i] = ((double)(dist[i]))/speed[i];
         }
         sort(time.begin(), time.end());
         double ready = 0.0;
